requirements.txt
Flask==3.0.0
flask-cors==4.0.0
tensorflow==2.18.0
Pillow==10.1.0
numpy==1.26.4
gunicorn==21.2.0
